+++ src/test/java/io/github/astrapi69/time/convert/OffsetDateTimeExtensionsTest.java
@@ -25,7 +25,6 @@
 package io.github.astrapi69.time.convert;
 
 import io.github.astrapi69.date.CreateDateExtensions;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
@@ -90,49 +89,6 @@ public class OffsetDateTimeExtensionsTest
 		offsetDateTime = OffsetDateTimeExtensions.toOffsetDateTime(localDateTime, zoneId);
 		actual = OffsetDateTimeExtensions.toZonedDateTime(offsetDateTime);
 		expected = ZonedDateTime.parse("2000-09-01T00:00+02:00");
-		assertEquals(expected, actual);
-	}
-
-	/**
-	 * Test method for {@link OffsetDateTimeExtensions#toOffsetDateTime(Date, ZoneId)
-	 */
-	@Test
-	@Disabled
-	// not working on github-actions
-	void testToOffsetDateTimeWithDateAndZoneId()
-	{
-
-		OffsetDateTime actual;
-		OffsetDateTime expected;
-		ZoneId zoneId;
-		Date date;
-
-		date = CreateDateExtensions.newDate(2000, 9, 1, 0, 0, 0);
-		zoneId = ZoneId.of("Europe/Paris");
-
-		actual = OffsetDateTimeExtensions.toOffsetDateTime(date, zoneId);
-		expected = OffsetDateTime.parse("2000-09-01T00:00+02:00");
-		assertEquals(expected, actual);
-	}
-
-	/**
-	 * Test method for {@link OffsetDateTimeExtensions#toOffsetDateTime(Date, ZoneOffset)
-	 */
-	@Test
-	@Disabled
-	// not working on github-actions
-	void testToOffsetDateTimeWithDateAndZoneOffset()
-	{
-		OffsetDateTime actual;
-		OffsetDateTime expected;
-		ZoneOffset zoneOffSet;
-		Date date;
-
-		date = CreateDateExtensions.newDate(2000, 9, 1, 0, 0, 0);
-		zoneOffSet = ZoneOffset.of("+02:00");
-
-		actual = OffsetDateTimeExtensions.toOffsetDateTime(date, zoneOffSet);
-		expected = OffsetDateTime.parse("2000-09-01T00:00+02:00");
 		assertEquals(expected, actual);
 	}